Many files: [[Lint]]
* util/hist.c:
* ntptrace/ntptrace.c:
* ntpq/ntpq.c:
* ntpdc/ntpdc.c:
* ntpdate/ntptimeset.c:
* ntpdate/ntpdate.c:
* ntpd/refclock_parse.c:
* ntpd/refclock_msfees.c:
* ntpd/refclock_jupiter.c:
* ntpd/ntp_refclock.c:
* ntpd/ntp_io.c:
* libparse/clk_wharton.c:
* libparse/clk_varitext.c:
* libparse/clk_trimtaip.c:
* libparse/clk_schmid.c:
* libparse/clk_rcc8000.c:
* libparse/clk_rawdcf.c:
* libparse/clk_meinberg.c:
* libparse/clk_hopf6021.c:
* libparse/clk_dcf7000.c:
* libparse/clk_computime.c:  Lint.

+++ ntpd/ntp_io.c
@@ -1161,7 +1161,7 @@ sendpkt(
         err = io_completion_port_sendto(inter, pkt, len, dest);
 	if (err != ERROR_SUCCESS)
 #else
-	cc = sendto(inter->fd, (char *)pkt, len, 0, (struct sockaddr *)dest,
+	cc = sendto(inter->fd, (char *)pkt, (size_t)len, 0, (struct sockaddr *)dest,
 		    sizeof(struct sockaddr_in));
 	if (cc == -1)
 #endif
